+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2021-12-13
+### Added
+- Medications Model
+- Medications#administration
+
 ## [1.7.4] - 2021-12-2
 ### Added
 - Component Model ID Property Changed
@@ -178,6 +183,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial Release
 
+[1.8.0]: https://github.com/WeInfuse/redox/compare/v1.7.4...v1.8.0
 [1.7.4]: https://github.com/WeInfuse/redox/compare/v1.7.3...v1.7.4
 [1.7.3]: https://github.com/WeInfuse/redox/compare/v1.7.2...v1.7.3
 [1.7.2]: https://github.com/WeInfuse/redox/compare/v1.7.1...v1.7.2